+++ json_serializable/test/src/checked_test_input.dart
@@ -3,13 +3,13 @@ part of '_json_serializable_test_input.dart';
 @ShouldGenerate(r'''
 WithANonCtorGetterChecked _$WithANonCtorGetterCheckedFromJson(
     Map<String, dynamic> json) {
-  return $checkedNew('WithANonCtorGetterChecked', json, () {
+  return $checkedCreate('WithANonCtorGetterChecked', json, ($checkedConvert) {
     $checkKeys(json,
         allowedKeys: const ['items'],
         requiredKeys: const ['items'],
         disallowNullValues: const ['items']);
     final val = WithANonCtorGetterChecked(
-      $checkedConvert(json, 'items',
+      $checkedConvert('items',
           (v) => (v as List<dynamic>).map((e) => e as String).toList()),
     );
     return val;